Move metron config into api package
