+++ packages/metron_agent_windows/spec
@@ -41,7 +41,6 @@ files:
 - loggregator/src/metron/api/*.go # gosub
 - loggregator/src/metron/clientpool/v1/*.go # gosub
 - loggregator/src/metron/clientpool/v2/*.go # gosub
-- loggregator/src/metron/config/*.go # gosub
 - loggregator/src/metron/egress/*.go # gosub
 - loggregator/src/metron/eventwriter/*.go # gosub
 - loggregator/src/metron/ingress/*.go # gosub
